Added random order sorting for inputs
Now the data that is passed to the class can be in any order. The data
will be sorted in descending order based on the number of cells. Any
other list will be sorted in the same way, thus, data can be added
dynamically while the ordering will always be ensured to be correct.