Added new safety factor policy for GCI calculation
The safety factor Fs used to be 1.25 for all grids. This has been
changed to conform with Roache where only two grids should use a safety
factor Fs=3 and three or more grids should use Fs=1.25
This has been implemented and the safety factor now dynamically changes
if the number of cells is changed.