Added a GCS class
The GCS class now looks after holding the data that the GCI object
produces. THis will be used to format the output in a user friendly
way, i.e. in form of a markdown or latex table (still to be
implemented). It also exposes a getter function which provides the user
access to the raw data should that be required.