Added Oberkampf and Roy correction
Now we can also use the correction proposed by Oberkampf and Roy, where
we limit the apparent order calculation to the simulation order should
it be larger than that. Furthermore, the safety factor is adapted based
on the discrepancy between the apparent and simulation order.

The correction is set to False by default and needs to be explicitly
switched on through the 'oberkampf_correction' flag in the constructor
call to either the GCI or GCS class. An example has been added to the
README file showing how to use it.